Finished barebones functionality of the API. Might have some methods missing that i forgot to add.

+++ AcheASaida/Contracts/InfoLabirintoDto.cs
@@ -7,4 +7,4 @@ public record InfoLabirintoDto(
     string Dificuldade,
     bool Completo,
     int Passos,
-    float PorcentagemExploracao);
+    decimal PorcentagemExploracao);
